+++ plugins/notifications-backend/src/index.ts
@@ -1,4 +1,0 @@
-export * from './service/router';
-export * from './service/permissions';
-export * from './plugin';
-export * from './dynamic';