+++ plugins/notifications-backend/src/service/constants.ts
@@ -1,6 +1,0 @@
-export const DefaultServiceUser = 'default/guest';
-export const DefaultMessageScope = 'user';
-export const DefaultPageNumber = 1;
-export const DefaultPageSize = 20;
-export const DefaultOrderBy = 'created';
-export const DefaultOrderDirection = 'desc';