+++ plugins/notifications/src/components/SystemNotifications/SystemNotifications.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import { NotificationsTable } from '../NotificationsTable';
-
-export const SystemNotifications = () => (
-  <NotificationsTable messageScope="system" title="Updates" />
-);