+++ plugins/notifications/src/constants.ts
@@ -1,2 +1,0 @@
-export const DebounceDelayMs = 1000;
-export const DefaultPollingIntervalMs = 5 * 1000;