+++ plugins/notifications/src/openapi/models/index.ts
@@ -1,9 +1,0 @@
-// @ts-nocheck
-/* tslint:disable */
-/* eslint-disable */
-export * from './Action';
-export * from './CreateBody';
-export * from './CreateBodyActionsInner';
-export * from './CreateNotification200Response';
-export * from './GetNotificationsCount200Response';
-export * from './Notification';