+++ plugins/notifications/src/routes.ts
@@ -1,5 +1,0 @@
-import { createRouteRef } from '@backstage/core-plugin-api';
-
-export const notificationsRootRouteRef = createRouteRef({
-  id: 'notifications',
-});