+++ plugins/topology-common/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @janus-idp/backstage-plugin-topology-common [1.2.0](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-topology-common@1.1.0...@janus-idp/backstage-plugin-topology-common@1.2.0) (2024-07-18)
+
+
+### Features
+
+* **deps:** update to backstage 1.28 ([#1891](https://github.com/janus-idp/backstage-plugins/issues/1891)) ([1ba1108](https://github.com/janus-idp/backstage-plugins/commit/1ba11088e0de60e90d138944267b83600dc446e5))
+
 ## @janus-idp/backstage-plugin-topology-common [1.1.0](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-topology-common@1.0.2...@janus-idp/backstage-plugin-topology-common@1.1.0) (2024-06-13)